requirements.txt
fastapi
uvicorn
redis
pydantic
sqlalchemy
aiosqlite
httpx
python-dotenv
